+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/src/DatagridDropdownFilter.tsx
@@ -27,8 +27,8 @@ export default function DatagridDropdownFilter(props: DatagridDropdownFilterCont
 
     const alertMessageMultipleFilters = (
         <Alert bootstrapStyle="danger">
-            To use multiple filters you need to define a filter identification in the properties of Drop-down filter or
-            have a &quot;Boolean or Enumeration&quot; attribute available.
+            The Drop-down filter widget can&apos;t be used with the filters options you have selected. It requires a
+            &quot;Boolean or Enumeration&quot; attribute to be selected.
         </Alert>
     );
 
@@ -80,11 +80,13 @@ export default function DatagridDropdownFilter(props: DatagridDropdownFilterCont
                 return (
                     <FilterComponent
                         ariaLabel={props.ariaLabel?.value}
+                        className={props.class}
                         defaultValue={defaultValues ?? props.defaultValue?.value}
                         emptyOptionCaption={props.emptyOptionCaption?.value}
                         multiSelect={props.multiSelect}
                         name={props.name}
                         options={options}
+                        styles={props.style}
                         tabIndex={props.tabIndex}
                         updateFilters={(values: FilterOption[]): void => {
                             const conditions = attributes