+++ packages/pluggableWidgets/datagrid-number-filter-web/src/DatagridNumberFilter.editorPreview.tsx
@@ -1,14 +1,21 @@
 import { createElement, ReactElement } from "react";
 import { DatagridNumberFilterPreviewProps } from "../typings/DatagridNumberFilterProps";
 import { FilterComponent } from "./components/FilterComponent";
+import { parseStyle } from "@mendix/piw-utils-internal";
 
-export function preview(props: DatagridNumberFilterPreviewProps): ReactElement {
+interface PreviewProps extends Omit<DatagridNumberFilterPreviewProps, "class"> {
+    className: string;
+}
+
+export function preview(props: PreviewProps): ReactElement {
     return (
         <FilterComponent
             adjustable={props.adjustable}
+            className={props.className}
             defaultFilter={props.defaultFilter}
             delay={props.delay ?? 500}
             placeholder={props.placeholder}
+            styles={parseStyle(props.style)}
         />
     );
 }