+++ packages/pluggableWidgets/datagrid-text-filter-web/src/components/FilterComponent.tsx
@@ -1,4 +1,4 @@
-import { createElement, ReactElement, useCallback, useEffect, useRef, useState } from "react";
+import { createElement, CSSProperties, ReactElement, useCallback, useEffect, useRef, useState } from "react";
 import { FilterSelector } from "@mendix/piw-utils-internal/components/web";
 import { debounce } from "@mendix/piw-utils-internal";
 
@@ -7,6 +7,7 @@ import classNames from "classnames";
 
 interface FilterComponentProps {
     adjustable: boolean;
+    className?: string;
     defaultFilter: DefaultFilterEnum;
     delay: number;
     name?: string;
@@ -14,6 +15,7 @@ interface FilterComponentProps {
     tabIndex?: number;
     screenReaderButtonCaption?: string;
     screenReaderInputCaption?: string;
+    styles?: CSSProperties;
     updateFilters?: (value: string, type: DefaultFilterEnum) => void;
     value?: string;
 }
@@ -47,7 +49,11 @@ export function FilterComponent(props: FilterComponentProps): ReactElement {
     }, [inputRef]);
 
     return (
-        <div className="filter-container" data-focusindex={props.tabIndex ?? 0}>
+        <div
+            className={classNames("filter-container", props.className)}
+            data-focusindex={props.tabIndex ?? 0}
+            style={props.styles}
+        >
             {props.adjustable && (
                 <FilterSelector
                     ariaLabel={props.screenReaderButtonCaption}
@@ -55,8 +61,13 @@ export function FilterComponent(props: FilterComponentProps): ReactElement {
                     defaultFilter={props.defaultFilter}
                     onChange={useCallback(
                         type => {
-                            setType(type);
-                            focusInput();
+                            setType(prev => {
+                                if (prev === type) {
+                                    return prev;
+                                }
+                                focusInput();
+                                return type;
+                            });
                         },
                         [focusInput]
                     )}