+++ scripts/release/createNativeModules.js
@@ -24,58 +24,65 @@ async function main() {
 
 async function createNMRModule() {
     console.log("Creating the Native Mobile Resource module.");
-    const pkgPath = join(process.cwd(), "packages/jsActions/mobile-resources-native/package.json");
+    const NMRFolder = join(process.cwd(), "packages/jsActions/mobile-resources-native");
     const tmpFolder = join(process.cwd(), "tmp/mobile-resources-native");
     const widgetFolders = await readdir(join(process.cwd(), "packages/pluggableWidgets"));
     const nativeWidgetFolders = widgetFolders
         .filter(folder => folder.includes("-native"))
         .map(folder => join(process.cwd(), "packages/pluggableWidgets", folder));
-    const {
-        name,
-        moduleName,
-        version,
-        marketplace: { minimumMXVersion },
-        testProject: { githubUrl, branchName },
-        repository: { url }
-    } = require(pkgPath);
+
+    const moduleInfo = getPackageInfo(NMRFolder);
+    await bumpVersionInPackageJson(NMRFolder, moduleInfo.version);
     await setLocalGitCredentials();
     await execShellCommand(
-        `git remote set-url origin https://${process.env.GH_USERNAME}:${process.env.GH_PAT}@${url.replace(
+        `git remote set-url origin https://${process.env.GH_USERNAME}:${process.env.GH_PAT}@${moduleInfo.url.replace(
             "https://",
             ""
         )}`
     );
     await execShellCommand(`echo "${process.env.GH_PAT}" | gh auth login --with-token`);
 
-    const changelog = await updateChangelogs(nativeWidgetFolders, pkgPath, version, moduleName, name);
-    await updateTestProject(tmpFolder, nativeWidgetFolders, githubUrl);
+    const changelog = await updateChangelogs(nativeWidgetFolders, moduleInfo);
+    await updateTestProject(tmpFolder, nativeWidgetFolders, moduleInfo.testProjectUrl);
 
     console.log("Creating module MPK..");
-    await createMxBuildContainer(tmpFolder, "NativeMobileResources", minimumMXVersion);
+    await createMxBuildContainer(tmpFolder, "NativeMobileResources", moduleInfo.minimumMXVersion);
     const mpkOutput = (await getFiles(tmpFolder, [`.mpk`]))[0];
 
-    console.log(`Creating Github release for module ${moduleName}`);
+    console.log(`Creating Github release for module ${moduleInfo.nameWithSpace}`);
     await execShellCommand(
-        `gh release create --title "${moduleName} ${version} - Mendix ${minimumMXVersion}" --notes "${changelog}" "${process.env.TAG}" "${mpkOutput}"`
+        `gh release create --title "${moduleInfo.nameWithSpace} ${moduleInfo.version} - Mendix ${moduleInfo.minimumMXVersion}" --notes "${changelog}" "${process.env.TAG}" "${mpkOutput}"`
     );
     await execShellCommand(`rm -rf ${tmpFolder}`);
     console.log("Done.");
 }
 
 // Update changelogs and create PR in widget-resources
-async function updateChangelogs(nativeWidgetFolders, pkgPath, version, moduleName, name) {
+async function updateChangelogs(nativeWidgetFolders, moduleInfo) {
     console.log("Updating changelogs..");
-    const moduleChangelogs = await getUnreleasedChangelogs(pkgPath, version);
+    const moduleChangelogs = await getUnreleasedChangelogs(moduleInfo);
     let nativeWidgetsChangelogs = "";
     for await (const folder of nativeWidgetFolders) {
-        nativeWidgetsChangelogs += (await combineWidgetChangelogs(folder)) || "";
+        const widgetInfo = getPackageInfo(folder);
+        const widgetChangelogs = await getUnreleasedChangelogs(widgetInfo);
+        nativeWidgetsChangelogs += changelogs || "";
+        if (widgetChangelogs) {
+            console.log(`Writing "${widgetInfo.nameWithSpace}" changelogs to ${widgetInfo.changelogPath}`);
+            await writeToChangelogs(widgetChangelogs, widgetInfo.changelogPath, widgetInfo.version);
+        }
     }
-    let changelog = moduleChangelogs ? `## [${version}] ${moduleName}\n${moduleChangelogs}\n` : "";
-    changelog = nativeWidgetsChangelogs ? `\n${nativeWidgetsChangelogs}` : changelog;
+    let changelog = moduleChangelogs
+        ? `## [${moduleInfo.version}] ${moduleInfo.nameWithSpace}\n${moduleChangelogs}\n`
+        : "";
+    changelog = nativeWidgetsChangelogs ? `${changelog}\n\n${nativeWidgetsChangelogs}` : changelog;
+    if (changelog) {
+        console.log(`Writing "${moduleInfo.nameWithSpace}" changelogs to ${moduleInfo.changelogPath}`);
+        await writeToChangelogs(changelog, moduleInfo.changelogPath, moduleInfo.version);
+    }
 
-    const changelogBranchName = `${name}-release-${version}`;
+    const changelogBranchName = `${moduleInfo.nameWithDash}-release-${moduleInfo.version}`;
     await execShellCommand(
-        `git checkout -b ${changelogBranchName} && git add . && git commit -m "chore(${name}): update changelogs" && git push --set-upstream origin ${changelogBranchName}`
+        `git checkout -b ${changelogBranchName} && git add . && git commit -m "chore(${moduleInfo.nameWithDash}): update changelogs" && git push --set-upstream origin ${changelogBranchName}`
     );
     await execShellCommand(
         `gh pr create --title "Updating all the changelogs" --body "This is an automated PR." --base master --head ${changelogBranchName}`
@@ -84,40 +91,69 @@ async function updateChangelogs(nativeWidgetFolders, pkgPath, version, moduleNam
     return changelog;
 }
 
-async function combineWidgetChangelogs(currentFolder) {
-    const { widgetName, version } = require(`${currentFolder}/package.json`);
-    const changelogPath = `${currentFolder}/CHANGELOG.md`;
+function getPackageInfo(path) {
+    const {
+        name,
+        widgetName,
+        moduleName,
+        version,
+        marketplace: { minimumMXVersion },
+        testProject: { githubUrl, branchName },
+        repository: { url }
+    } = require(`${path}/package.json`);
+    return {
+        nameWithDash: name,
+        nameWithSpace: moduleName ?? widgetName,
+        version,
+        minimumMXVersion,
+        url,
+        testProjectUrl: githubUrl,
+        testProjectBranchName: branchName,
+        changelogPath: `${path}/CHANGELOG.md`
+    };
+}
+
+async function getUnreleasedChangelogs({ name, version, changelogPath }) {
     try {
         await access(changelogPath);
-        const changelogs = await getUnreleasedChangelogs(changelogPath, version);
+        const content = await readFile(changelogPath, "utf8");
+        const changelogs = content.match(/(?<=## \[unreleased\]\n)((?!## \[\d+\.\d+\.\d+\])\W|\w)*/i)?.[0].trim();
+        const releasedVersions = content.match(/(?<=## \[)\d+\.\d+\.\d+(?=\])/g);
+        if (releasedVersions?.includes(version)) {
+            throw new Error(
+                `It looks like version ${version} from package.json is already released. Did you forget to bump the version?`
+            );
+        }
+
         return changelogs
-            ? `## [${version}] ${widgetName}
+            ? `## [${version}] ${name}
             
-            ${changelogs}`
+            ${changelogs}\n\n`
             : "";
     } catch (error) {
         console.warn(`${changelogPath} does not exist.`);
     }
 }
 
-async function getUnreleasedChangelogs(changelogFile, version) {
-    const content = await readFile(changelogFile, "utf8");
-    const unreleasedChangelogs = content.match(/(?<=## \[unreleased\]\n)((?!## \[\d+\.\d+\.\d+\])\W|\w)*/i)?.[0].trim();
-    const releasedVersions = content.match(/(?<=## \[)\d+\.\d+\.\d+(?=\])/g);
-    if (releasedVersions?.includes(version)) {
+async function writeToChangelogs(changelogs, changelogPath, version) {
+    const d = new Date();
+    const date = `${d.getFullYear()}-${d.getMonth() + 1}-${d.getDate()}`;
+    const newContent = changelogs.replace(`## [Unreleased]`, `## [Unreleased]\n\n## [${version}] ${date}`);
+    await writeFile(changelogPath, newContent);
+}
+
+async function bumpVersionInPackageJson(NMRFolder, moduleVersionOld) {
+    const moduleVersionNew = process.env.TAG.split("-v")[1];
+    if (moduleVersionOld === moduleVersionNew) {
         throw new Error(
-            `It looks like version ${version} from package.json is already released. Did you forget to bump the version?`
+            `It looks like version ${moduleVersionOld} of "${moduleInfo.nameWithSpace}" has been released already. Did you manually bump the version without releasing? Then you should revert the bump before retrying.`
         );
+    } else {
+        console.log(`Bumping "${moduleInfo.nameWithSpace}" version from ${moduleVersionOld} to ${moduleVersionNew}..`);
+        const pkg = require(join(NMRFolder, "package.json"));
+        pkg.version = moduleVersionNew;
+        await writeFile(pkgPath, pkg);
     }
-
-    if (unreleasedChangelogs) {
-        const d = new Date();
-        const date = `${d.getFullYear()}-${d.getMonth() + 1}-${d.getDate()}`;
-        const newContent = content.replace(`## [Unreleased]`, `## [Unreleased]\n\n## [${version}] ${date}`);
-        await writeFile(changelogFile, newContent);
-    }
-
-    return unreleasedChangelogs;
 }
 
 // Update test project with latest changes
@@ -217,5 +253,3 @@ async function setLocalGitCredentials(workingDirectory) {
     await execShellCommand(`git config user.name "${process.env.GH_NAME}"`, workingDirectory);
     await execShellCommand(`git config user.email "${process.env.GH_EMAIL}"`, workingDirectory);
 }
-
-// TODO: add functionality to bump NMR module version.