+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
-- We added new options to allow filtering in multiple attributes. To enable it make sure you have "Advanced" and "Show header filters" on.
+- We added new options to allow filtering in multiple attributes. To enable it make sure you have "Show header filters" on.
 - We added a toolbox category and toolbox tile image for Studio & Studio Pro.
+
+### Removed
+- We removed the advanced toggle for Studio Pro (kept for Studio)
+
+### Changed
+- We fixed the positioning of the filters when virtual scrolling was enabled and there was no data being presented.
 
 ## [1.2.1] - 2021-07-01